do some other stuff i forgot
