--- src/render/tripOffer.js
@@ -1,5 +1,0 @@
-const makeTripPointOffer = ([name, price]) => `<li>
-  <button class="trip-point__offer">${name} +&euro;&nbsp;${price}</button>
-</li>`;
-
-export default makeTripPointOffer;
